Adaptations Python3 int()

+++ noethys/Ctrl/CTRL_Commande_repas.py
@@ -239,7 +239,7 @@ class CaseRenderer(GridCellRenderer):
                 dc.DrawRoundedRectangleRect(wx.Rect(xRond, yRond, largeurRond, hauteurRond), hauteurRond / 2.0)
 
             # Texte
-            xTexte = xRond + largeurRond / 2.0 - largeurTexte / 2.0
+            xTexte = int(xRond + largeurRond / 2.0 - largeurTexte / 2.0)
             yTexte = yRond + hauteurRond / 2.0 - hauteurTexte / 2.0
             dc.DrawText(texte, xTexte, yTexte)
